chore: Update Terraform backend configuration
Update the Terraform backend configuration in the manual-test-release.yml file. This commit updates the tf_vars_file and tf_key values to reflect the changes in the backend configuration. The tf_vars_file is changed to "testing.tfvars" and the tf_key is changed to "tf-mod-tests-sec-back". Additionally, the az_container_name is updated to "gh-secure-backend".

Co-authored-by: dependabot[bot] <support@github.com>